Add files via upload
A startup company that sells food products. Need to research user behavior for the company's application.

First, study the sales funnel. Find out how users reach the purchase stage. How many users actually reach this stage? How many get stuck in earlier stages? Which specific stages?

Then, take a look at the results of an A/A/B test (*keep reading to learn more about A/A/B test*). The design team would like to change the fonts throughout the entire app, but management fears that users may find the new design intimidating. For this reason, they've decided to make a decision based on the results of an A/A/B test.

Users are divided into **three groups**: two control groups get the old fonts, and one test group gets the new fonts. Determine which set of fonts produces better results.

Creating two A groups has certain advantages. We can establish the principle that we'll only trust the accuracy of our tests when both control groups are similar. If there are significant differences between the A groups, this may help us identify factors that could be skewing the results. Comparing control groups also tells us how much time and data we'll need when conducting future tests.

You'll use the same dataset for both the general analysis and the A/A/B test analysis. In real projects, experiments are carried out continuously. The analytics team studies app quality using general data, without paying attention to whether users are participating in experiments.